requirements.txt
black==21.5b1
celery==4.4.2
Django==3.2.2
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
flower==0.9.3
gunicorn==20.1.0
psycopg2-binary==2.8.6
python-dotenv==0.17.1
redis==3.5.3
scipy==1.6.3
